Refatorando interfaces
