--- api/Interface/ControllerInterface.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace Bifrost\Interface;
-
-/**
- * Interface ControllerInterface
- *
- * Interface para os controllers
- *
- * @package Bifrost\Rules
- * @author Felipe dos S. Cavalca
- */
-interface ControllerInterface {
-
-    public function index(); // Método padrão para o controller
-
-}
